Create Update-UserConfigAfterMigration.ps1
This is a script to find all users in a migration batch and then add them to multiple AD groups to set cache mode in Outlook, add the EXO licence and configure Intune, and add the Intune licence. The group name should be adjusted as necessary to match your own AD group name and configure the group in Azure to perform the appropriate actions for you. 
This is a custom group for custom actions after migration.